tutorial/requirement.txt
Django==2.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Pygments==2.2.0
PyMySQL==0.9.2
